requirements.txt
coqui-tts>=0.21.3
rvc>=0.3.5
flask>=3.1.0
flask-socketio
colorlog>=6.9.0
